+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ### 2026-04-20
+- feat: add player stats resource
 - feat: make camera follow player
 
 ### 2026-04-19